requirements.txt
pytesseract
python-docx
pdf2image